move chart title on top centre and add some styling
